+++ Tests/Command/CreateUserCommandTest.php
@@ -23,7 +23,6 @@ class CreateUserCommandTest extends \PHPUnit_Framework_TestCase
     {
         $commandTester = $this->createCommandTester($this->getContainer('user', 'pass', 'email', true, false));
         $exitCode = $commandTester->execute(array(
-            'command' => 'fos:user:create', // BC for SF <2.4 see https://github.com/symfony/symfony/pull/8626
             'username' => 'user',
             'email' => 'email',
             'password' => 'pass',
@@ -36,58 +35,8 @@ class CreateUserCommandTest extends \PHPUnit_Framework_TestCase
         $this->assertRegExp('/Created user user/', $commandTester->getDisplay());
     }
 
-    /**
-     * @group legacy
-     */
-    public function testExecuteInteractiveWithDialogHelper()
-    {
-        if (!class_exists('Symfony\Component\Console\Helper\DialogHelper')) {
-            $this->markTestSkipped('Using the DialogHelper is not possible on Symfony 3+.');
-        }
-
-        $application = new Application();
-
-        $dialog = $this->getMock('Symfony\Component\Console\Helper\DialogHelper', array(
-            'askAndValidate',
-            'askHiddenResponseAndValidate',
-        ));
-        $dialog->expects($this->at(0))
-            ->method('askAndValidate')
-            ->will($this->returnValue('user'));
-
-        $dialog->expects($this->at(1))
-            ->method('askAndValidate')
-            ->will($this->returnValue('email'));
-
-        $dialog->expects($this->at(2))
-            ->method('askHiddenResponseAndValidate')
-            ->will($this->returnValue('pass'));
-
-        $helperSet = new HelperSet(array(
-            'dialog' => $dialog,
-        ));
-        $application->setHelperSet($helperSet);
-
-        $commandTester = $this->createCommandTester(
-            $this->getContainer('user', 'pass', 'email', true, false), $application
-        );
-        $exitCode = $commandTester->execute(array(
-            'command' => 'fos:user:create', // BC for SF <2.4 see https://github.com/symfony/symfony/pull/8626
-        ), array(
-            'decorated' => false,
-            'interactive' => true,
-        ));
-
-        $this->assertEquals(0, $exitCode, 'Returns 0 in case of success');
-        $this->assertRegExp('/Created user user/', $commandTester->getDisplay());
-    }
-
     public function testExecuteInteractiveWithQuestionHelper()
     {
-        if (!class_exists('Symfony\Component\Console\Helper\QuestionHelper')) {
-            $this->markTestSkipped('The question helper not available.');
-        }
-
         $application = new Application();
 
         $helper = $this->getMock('Symfony\Component\Console\Helper\QuestionHelper', array(